+++ R/package-CovTools.R
@@ -14,45 +14,6 @@
 #' \code{\link{CovMean}} \tab compute mean covariance matrix
 #' }
 #'
-#' @section Statistical Methods:
-#' We provide statistical methods for
-#' (1) \strong{Covariance Matrix Estimation},
-#' \tabular{ll}{
-#' \emph{name of a function}\tab \emph{description} \cr
-#' \code{\link{CovEst.adaptive}} \tab Adaptive Thresholding. \cr
-#' \code{\link{CovEst.hard}} \tab Hard Thresholding. \cr
-#' \code{\link{CovEst.hardPD}} \tab Hard Thresholding under Positive-Definiteness Constraint. \cr
-#' \code{\link{CovEst.nearPD}} \tab Nearest Positive-Definite Matrix Projection. \cr
-#' \code{\link{CovEst.soft}} \tab Soft Thresholding.
-#' }
-#' (2) \strong{Precision Matrix Estimation}
-#' \tabular{ll}{
-#' \emph{name of a function}\tab \emph{description} \cr
-#' \code{\link{PreEst.2014An}} \tab Banded Precision Matrix Estimation via Bandwidth Test. \cr
-#' \code{\link{PreEst.2014Banerjee}} \tab Bayesian Estimation of a Banded Precision Matrix (Banerjee 2014). \cr
-#' \code{\link{PreEst.2017Lee}} \tab Bayesian Estimation of a Banded Precision Matrix (Lee 2017). \cr
-#' \code{\link{PreEst.glasso}} \tab Graphical Lasso.
-#' }
-#' (3) \strong{1-Sample Covariance Tests}
-#' \tabular{ll}{
-#' \emph{name of a function} \tab \emph{description} \cr
-#' \code{\link{BCovTest1.mxPBF}} \tab Bayesian Test using Maximum Pairwise Bayes Factor. \cr
-#' \code{\link{CovTest1.2013Cai}} \tab Test by Cai and Ma (2013). \cr
-#' \code{\link{CovTest1.2014Srivastava}} \tab Test by Srivastava, Yanagihara, and Kubokawa (2014).
-#' }
-#' (4) \strong{2-Sample Covariance Tests}
-#' \tabular{ll}{
-#' \emph{name of a function} \tab \emph{description} \cr
-#' \code{\link{CovTest2.2013Cai}} \tab Test by Cai and Ma (2013).
-#' }
-#' (5) \strong{1-Sample Diagonality Tests}
-#' \tabular{ll}{
-#' \emph{name of a function} \tab \emph{description} \cr
-#' \code{\link{BDiagTest1.mxPBF}} \tab Bayesian Test using Maximum Pairwise Bayes Factor. \cr
-#' \code{\link{DiagTest1.2011Cai}} \tab Test by Cai and Jiang (2011). \cr
-#' \code{\link{DiagTest1.2015Lan}} \tab Test by Lan et al. (2015).
-#' }
-#'
 #'
 #' @docType package
 #' @name package-CovTools