+++ NEWS.md
@@ -1,8 +1,8 @@
-# News for Package SAMCpack
+# News for Package CovTools
 ### Changes in version 0.5.2
+  * Function list is moved to README file.
   * Following functions are added,
-    - a
-    - b
+    - `CovEst.2003LW` : linear shrinkage covariance estimator.
     
 ### Changes in version 0.5.1
   * Functions for hypothesis tests are separated for clarity.